chore: Convert next.config.js to next.config.ts, as rest of repo is TS
- Remove remotePatterns, as data-stories in branding package now uses plain <img>
- Remove reactStrictMode: false from 3937109

Sentry:
- Gate on NEXT_PUBLIC_SENTRY_DSN_URL, instead of SENTRY_AUTH_TOKEN from f4753a8
- Set widenClientFileUpload based on SENTRY_AUTH_TOKEN being set (which it requires)
- Remove hideSourceMaps, as it was deprecated and removed from @sentry/core
  https://github.com/getsentry/sentry-javascript/releases/tag/8.43.0
- Remove automaticVercelMonitors, as site is not on Vercel
