--- next.config.js
@@ -1,84 +1,0 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-/** @type {import('next').NextConfig} */
-
-const { withSentryConfig } = require('@sentry/nextjs');
-const withNextIntl = require('next-intl/plugin')();
-
-const nextConfig = {
-  output: 'standalone',
-  transpilePackages: ['opub-ui', 'ids-drr-branding'],
-  serverExternalPackages: [
-    // Dependencies of @sentry/node.
-    '@prisma/instrumentation',
-    '@fastify/otel',
-  ],
-  images: {
-    remotePatterns: [
-      {
-        protocol: 'https',
-        hostname: 'miro.medium.com',
-        port: '',
-      },
-    ],
-  },
-  reactStrictMode: false,
-  // https://github.com/CivicDataLab/opub-mono/pull/403
-  webpack: (config) => {
-    config.module.rules.push({
-      test: /\.(js|mjs)$/,
-      enforce: 'pre',
-      include: /node_modules\/opub-ui/,
-      use: [
-        {
-          loader: 'source-map-loader',
-          options: { filterSourceMappingUrl: () => 'remove' },
-        },
-      ],
-    });
-    return config;
-  },
-};
-
-const wrappedConfig = withNextIntl(nextConfig);
-
-module.exports = process.env.SENTRY_AUTH_TOKEN
-  ? withSentryConfig(wrappedConfig, {
-      // For all available options, see:
-      // https://github.com/getsentry/sentry-webpack-plugin#options
-
-      org: process.env.SENTRY_ORG_NAME,
-      project: process.env.SENTRY_PROJECT_NAME,
-      sentryUrl: process.env.SENTRY_URL,
-
-      // Only print logs for uploading source maps in CI
-      silent: !process.env.CI,
-
-      // For all available options, see:
-      // https://docs.sentry.io/platforms/javascript/guides/nextjs/manual-setup/
-
-      // Upload a larger set of source maps for prettier stack traces (increases build time)
-      widenClientFileUpload: true,
-
-      // Uncomment to route browser requests to Sentry through a Next.js rewrite to circumvent ad-blockers.
-      // This can increase your server load as well as your hosting bill.
-      // Note: Check that the configured route will not match with your Next.js middleware, otherwise reporting of client-
-      // side errors will fail.
-      // tunnelRoute: "/monitoring",
-
-      // Hides source maps from generated client bundles
-      hideSourceMaps: true,
-
-      webpack: {
-        // Automatically tree-shake Sentry logger statements to reduce bundle size
-        treeshake: {
-          removeDebugLogging: true,
-        },
-
-        // Enables automatic instrumentation of Vercel Cron Monitors. (Does not yet work with App Router route handlers.)
-        // See the following for more information:
-        // https://docs.sentry.io/product/crons/
-        // https://vercel.com/docs/cron-jobs
-        automaticVercelMonitors: true,
-      },
-    })
-  : wrappedConfig;
